+++ packages/Router/CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+- cache-count implementation on domx-route
+### Updated
+- routing behavior - DEFINE
+
 ## [0.2.0] 2021-09-29
 ### Added
 - window event for not found routes